change warning to error
